take /components/search/Connector.tsx apart
